New updates. Moved files to espresso and updated imports and test.
