+++ espresso/src/test/java/com/progressive/kherkin/common/StepNavigatorTest.kt
@@ -1,10 +1,10 @@
 package com.progressive.kherkin.common
 
-import com.progressive.kherkin.common.testcore.Navigable
-import com.progressive.kherkin.common.testcore.PathSegment
 import com.progressive.kherkin.common.testcore.PreconditionsData
-import com.progressive.kherkin.common.testcore.StepNavigator
 import com.progressive.kherkin.common.testcore.TestLogger
+import com.progressive.kherkin.espresso.custom.Navigable
+import com.progressive.kherkin.espresso.custom.PathSegment
+import com.progressive.kherkin.espresso.custom.StepNavigator
 import org.junit.After
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertTrue
