+++ sampleapp/src/androidTest/java/com/progressive/sampleapp/screens/espresso/FinalScreen.kt
@@ -6,8 +6,8 @@ import com.progressive.kherkin.sampleapp.R
 import com.progressive.kherkin.common.screen.Screen
 import com.progressive.kherkin.common.screen.Trait
 import com.progressive.kherkin.espresso.steps.actions.ITouchButton
-import com.progressive.kherkin.common.testcore.Navigable
-import com.progressive.kherkin.common.testcore.PathSegment
+import com.progressive.kherkin.espresso.custom.Navigable
+import com.progressive.kherkin.espresso.custom.PathSegment
 import com.progressive.kherkin.common.testcore.When
 import com.progressive.sampleapp.activities.xml.FinalActivity
 import com.progressive.sampleapp.setup.SamplePreconditionsData
