+++ sampleapp/src/androidTest/java/com/progressive/sampleapp/screens/espresso/MainScreen.kt
@@ -8,8 +8,8 @@ import androidx.test.core.app.ApplicationProvider
 import com.progressive.kherkin.sampleapp.R
 import com.progressive.kherkin.common.screen.Screen
 import com.progressive.kherkin.common.screen.Trait
-import com.progressive.kherkin.common.testcore.Navigable
-import com.progressive.kherkin.common.testcore.PathSegment
+import com.progressive.kherkin.espresso.custom.Navigable
+import com.progressive.kherkin.espresso.custom.PathSegment
 import com.progressive.kherkin.compose.steps.testcore.ComposeNavigable
 import com.progressive.kherkin.compose.steps.testcore.ComposePathSegment
 import com.progressive.sampleapp.activities.xml.MainActivity
